requirements.txt
flask
pytesseract
pillow
openpyxl
gunicorn